Move FE tests to react sub folder
