+++ react-ui/test/components/addpin.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow, mount } from 'enzyme';
-import Signin from '../../react-ui/src/components/SignIn';
+import Signin from '../../src/components/SignIn';
 import { Provider } from 'react-redux';
 
 const setup = () => {
